requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
openai>=1.30.0
google-generativeai>=0.3.0
pandas==2.2.3
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
openpyxl==3.1.2
requests==2.31.0